Create Script.mjs
This is took me about a week, this isn't the final version of the program, this is jsut a prototype to see if the program itself would work. Feel free to try this code out yourself. I also have a node version of this program that can be run in a terminal based off of Javascript.